configure droplet with new ssh key
On every deploy, generate a new random SSH key and setup the
DigitalOcean droplet with the key in its authorized_keys config.

For now, write the associated private key to ~/dochaincore.pem.
Later, we can avoid writing the private key to the filesystem at
all. We can ssh in directly from within the same process, generate
a Chain Core client token using corectl, remove our public key from
the authorized_keys and close the connection.

This gives us a simple way to create a client token without relying
on the assumption that the user's DigitalOcean SSH keys are in the
ssh agent keyring.